+++ test/logger.ts
@@ -134,9 +134,7 @@ describe('getModifiedData', () => {
   it('simple text with error', () => {
     const modifiedData = getModifiedData(sampleText, undefined, true);
     const expectedOutput =
-      JSON.stringify(
-        Object.assign(JSON.parse(expectedTextOutput), {severity: 'ERROR'})
-      ) + '\n';
+      JSON.stringify(Object.assign(JSON.parse(expectedTextOutput))) + '\n';
     assert.equal(modifiedData, expectedOutput);
   });